Add first version for ROS 2 humble
The configuration here provided uses the latest stable Neovim version
as well as the latest version of a certain collection of plugins.

The plugins here selected can speed up the initial setup of Neovim and
allow you to quickly and efficiently edit, compile, and debug C++ and
Python code in the context of ROS 2.

For validating the configuration a Dockerfile where the config is
deployed into a clean environment.